Remove unused refreshPipeline var.
