+++ hadoop-ozone/common/src/main/java/org/apache/hadoop/ozone/OmUtils.java
@@ -273,11 +273,11 @@ public final class OmUtils {
     case TenantListUser:
     case EchoRPC:
     case RangerBGSync:
-    case GetKeyInfo:
       // RangerBGSync is a read operation in the sense that it doesn't directly
       // write to OM DB. And therefore it doesn't need a OMClientRequest.
       // Although indirectly the Ranger sync service task could invoke write
       // operation SetRangerServiceVersion.
+    case GetKeyInfo:
       return true;
     case CreateVolume:
     case SetVolumeProperty:
