+++ hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/ozone/TestStorageContainerManagerHelper.java
@@ -89,7 +89,6 @@ public class TestStorageContainerManagerHelper {
           .setVolumeName(bucket.getVolumeName())
           .setBucketName(bucket.getName())
           .setKeyName(key)
-          .setRefreshPipeline(true)
           .build();
       OmKeyInfo location = cluster.getOzoneManager()
           .lookupKey(arg);
