+++ hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/ozone/scm/pipeline/TestSCMPipelineBytesWrittenMetrics.java
@@ -100,7 +100,7 @@ public class TestSCMPipelineBytesWrittenMetrics {
 
     OmKeyArgs.Builder builder = new OmKeyArgs.Builder();
     builder.setVolumeName(volumeName).setBucketName(bucketName)
-        .setKeyName(keyName).setRefreshPipeline(true);
+        .setKeyName(keyName);
 
     OzoneKeyDetails keyDetails = bucket.getKey(keyName);
     Assertions.assertEquals(keyName, keyDetails.getName());
