+++ hadoop-ozone/ozonefs-common/src/main/java/org/apache/hadoop/fs/ozone/OzoneClientUtils.java
@@ -210,7 +210,7 @@ public final class OzoneClientUtils {
     }
     OmKeyArgs keyArgs = new OmKeyArgs.Builder().setVolumeName(volume.getName())
         .setBucketName(bucket.getName()).setKeyName(keyName)
-        .setRefreshPipeline(true).setSortDatanodesInPipeline(true)
+        .setSortDatanodesInPipeline(true)
         .setLatestVersionLocation(true).build();
     OmKeyInfo keyInfo = rpcClient.getOzoneManagerClient().lookupKey(keyArgs);
